fix: latidude and longitude can be set (closes #373)
This fixes the inverted NaN-check to see if coordinates are valid, and makes the logic for choosing coodrinates or defaults a bit more robust. They can now fail independently of each other, so that an error in lat doesn't prevent long from being set.

The zh_CN translations still need to be adjusted slightly, for now I'm still usign the old warning that didn't differntiate between lat/long.